progress admin posts page

+++ includes/delete_posts.php
@@ -4,6 +4,10 @@ session_start();
 
 include ("database_connection.php");
 
+// Check connection
+if ($pdo->connect_error) {
+    die("Connection failed: " . $pdo->connect_error);
+} 
 
 if(isset($_GET['post_id'])){
     $statement = $pdo->prepare("DELETE FROM posts WHERE post_id = :post_id");
